docbook-xsl: slightly modernise
Instead of installing a subset of docbook-xsl, remove the tools (23MB) and ship
everything else.  This ensure we don't accidentally forget to install a useful
file (such as the provided catalog).

Instead of manually writing XML catalog lines (four entries for docbook-xsl),
simply chain to the docbook-xsl catalog.

The same cleanups can be done for the rest of the components being shipped, but
that can be done at a later date.